Feat: README

+++ src/solver/Solver.java
@@ -12,5 +12,5 @@ public interface Solver {
     public String getSource();
     public String getTarget();
     public Integer getTotalNodesVisited();
-    public Long getSolveTime();
+    public Double getSolveTime();
 }
